% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{DEFAULT_BASE_URL}
\alias{DEFAULT_BASE_URL}
\title{The address of the public Quicopt service}
\usage{
DEFAULT_BASE_URL
}
\value{
Not a function but a constant: a character string, the address of
the public service.
}
\description{
\code{\link[=solve]{solve()}} and \code{\link[=submit]{submit()}} send models to this address unless they are given
another \code{base_url}.
}
\examples{
DEFAULT_BASE_URL
}
