% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{index_set}
\alias{index_set}
\title{A named index set with concrete elements}
\usage{
index_set(name, elements)
}
\arguments{
\item{name}{The set's name.}

\item{elements}{A list of integers and strings.}
}
\description{
A named index set with concrete elements
}
